backend/requirements.txt
 fastapi
  uvicorn[standard]
  python-multipart
  sqlalchemy
  alembic
  anthropic
  httpx
  aiohttp
  python-dotenv
  pydantic
  pydantic-settings
  aiofiles
  requests
  PyYAML
  click
  pandas
  sseclient-py
  google-api-python-client
  google-auth
  google-auth-httplib2